% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphdata.R
\name{graph_data}
\alias{graph_data}
\title{Graph Data}
\usage{
graph_data(data = df, id = "user_id", product_id = "product_id")
}
\arguments{
\item{data}{dataframe of purchase history}

\item{id}{column name identifying the user id}

\item{product_id}{column name identifying the id of the objects purchased}
}
\value{
an object of class \code{"iGraph"}
}
\description{
Creates an object of class \code{"iGraph"} by transforming line order data into network data.
}
